Some organization + renaming.
